form collection data logic 90% completed
