+++ src/components/organisms/index.js
@@ -1,1 +1,2 @@
 export * from "./header";
+export * from "./FormSection";
